Mobile version - edited
Hey Yana! Thank you for your comments. I totally agree with your opinion on the images. I was trying something different and it worked, so I left it the way it was. Thank you for explaing me the reason why I should use the img in html. :)) I edited the version and hope, you will like this one better :))
thank you again!